backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
pydantic-settings>=2.1.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pyjwt>=2.8.0
redis>=5.0.1
python-multipart>=0.0.6
qrcode[pil]
pillow
cryptography
httpx
email-validator
boto3
